+++ blocklets/ai-studio/src/components/file-editor/input/InputTable.tsx
@@ -149,7 +149,7 @@ export default function InputTable({
     return true;
   };
 
-  const parameters = sortBy(Object.values(assistant.parameters ?? {}), (i) => i.index);
+  const parameters = sortBy(Object.values(assistant.parameters ?? {}), (i) => i.index).filter((i) => !i.data.hidden);
   const { data: knowledge = [] } = useRequest(() => getDatasets({ projectId }));
 
   const FROM_MAP = useMemo(() => {
@@ -195,7 +195,12 @@ export default function InputTable({
             };
 
             return (
-              <Box height={33} display="flex" alignItems="center" gap={0.5}>
+              <Box
+                height={33}
+                display="flex"
+                alignItems="center"
+                gap={0.5}
+                sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}>
                 <Box component={Icon} icon={iconMap[parameter.key]} fontSize={16} />
                 <Box>{parameter.key}</Box>
               </Box>
@@ -203,7 +208,12 @@ export default function InputTable({
           }
 
           return (
-            <Stack direction="row" alignItems="center" gap={0.5} minWidth={100}>
+            <Stack
+              direction="row"
+              alignItems="center"
+              gap={0.5}
+              minWidth={100}
+              sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}>
               <Box component={Icon} icon={FormsIcon} fontSize={16} />
 
               <WithAwareness
@@ -212,9 +222,10 @@ export default function InputTable({
                 sx={{ top: 4, right: -8 }}
                 path={[assistant.id, 'parameters', parameter?.id ?? '', 'key']}>
                 <Input
+                  sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}
                   id={`${parameter.id}-key`}
                   fullWidth
-                  readOnly={readOnly}
+                  readOnly={readOnly || parameter.hidden}
                   placeholder={t('inputParameterKeyPlaceholder')}
                   value={parameter.key || ''}
                   onChange={(e) => {
@@ -236,7 +247,7 @@ export default function InputTable({
         flex: 1,
         renderCell: ({ row: { data: parameter } }) => {
           if (parameter.type === 'source' && parameter.source?.variableFrom === 'history') {
-            return <Box>{t('history.title')}</Box>;
+            return <Box sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}>{t('history.title')}</Box>;
           }
 
           return (
@@ -266,7 +277,10 @@ export default function InputTable({
             const { source } = parameter;
             if (source.variableFrom === 'tool') {
               return (
-                <Stack direction="row" alignItems="center">
+                <Stack
+                  direction="row"
+                  alignItems="center"
+                  sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}>
                   <ListItemIcon sx={{ minWidth: 20 }}>
                     <Icon icon={BracesIcon} />
                   </ListItemIcon>
@@ -282,7 +296,10 @@ export default function InputTable({
                 (x) => x.key === source.variable?.key && x.scope && source.variable.scope
               );
               return (
-                <Stack direction="row" alignItems="center">
+                <Stack
+                  direction="row"
+                  alignItems="center"
+                  sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}>
                   {variable?.type?.type ? (
                     <>
                       <ListItemIcon sx={{ minWidth: 20 }}>{TYPE_ICON_MAP[variable.type.type]}</ListItemIcon>
@@ -297,7 +314,10 @@ export default function InputTable({
 
             if (parameter.source.variableFrom === 'knowledge') {
               return (
-                <Stack direction="row" alignItems="center">
+                <Stack
+                  direction="row"
+                  alignItems="center"
+                  sx={{ color: parameter.hidden ? 'text.disabled' : undefined }}>
                   <ListItemIcon sx={{ minWidth: 20 }}>
                     <Icon icon={CursorTextIcon} />
                   </ListItemIcon>
@@ -386,6 +406,10 @@ export default function InputTable({
               const idReferenced = checkVariableReferenced(parameter.id, parameter.key || '');
 
               const getBackgroundColor = (theme: Theme) => {
+                if (parameter.hidden) {
+                  return alpha(theme.palette.common.black, theme.palette.action.hoverOpacity);
+                }
+
                 if (!idReferenced) {
                   return alpha(theme.palette.warning.light, theme.palette.action.focusOpacity);
                 }
@@ -398,6 +422,10 @@ export default function InputTable({
               };
 
               const getHoverBackgroundColor = (theme: Theme) => {
+                if (parameter.hidden) {
+                  return alpha(theme.palette.common.black, theme.palette.action.hoverOpacity);
+                }
+
                 if (!idReferenced) {
                   return `${alpha(theme.palette.warning.light, theme.palette.action.selectedOpacity)} !important`;
                 }
@@ -405,8 +433,20 @@ export default function InputTable({
                 return 'grey.100';
               };
 
+              const title = () => {
+                if (parameter.hidden) {
+                  return undefined;
+                }
+
+                if (idReferenced) {
+                  return undefined;
+                }
+
+                return t('variableNotReferenced');
+              };
+
               return (
-                <Tooltip title={idReferenced ? undefined : t('variableNotReferenced')} placement="bottom-start">
+                <Tooltip title={title()} placement="bottom-start">
                   <TableRow
                     key={parameter.id}
                     ref={(ref) => {
@@ -433,11 +473,7 @@ export default function InputTable({
                           <TableCell
                             key={column.field}
                             align={column.align}
-                            sx={{
-                              position: 'relative',
-                              px: 0,
-                              ...getDiffBackground('parameters', parameter.id),
-                            }}>
+                            sx={{ position: 'relative', px: 0, ...getDiffBackground('parameters', parameter.id) }}>
                             {index === 0 && (
                               <Stack
                                 className="hover-visible center"
@@ -543,8 +579,19 @@ function SelectFromSource({
     <>
       <PopperMenu
         ref={ref}
-        BoxProps={{
-          sx: { my: 1, p: 0, cursor: 'pointer' },
+        ButtonProps={{
+          variant: 'text',
+          sx: {
+            my: 1,
+            p: 0,
+            cursor: 'pointer',
+            color: parameter.hidden ? 'text.disabled' : 'text.primary',
+            fontWeight: 400,
+            ':hover': {
+              backgroundColor: 'transparent',
+            },
+          },
+          disabled: parameter.hidden,
           children: (
             <Box>
               <Box className="center" gap={1} justifyContent="flex-start">
@@ -664,7 +711,8 @@ function SelectInputType({
           disabled={
             parameter.key === 'question' ||
             parameter.from === FROM_PARAMETER ||
-            parameter.from === FROM_KNOWLEDGE_PARAMETER
+            parameter.from === FROM_KNOWLEDGE_PARAMETER ||
+            parameter.hidden
           }
           variant="standard"
           hiddenLabel
@@ -1131,6 +1179,10 @@ function checkKeyParameterIsUsed({ value, key }: { value: AssistantYjs; key: str
   }
 
   const parameters = Object.values(value?.parameters || {}).flatMap((x) => {
+    if (x.data.hidden) {
+      return [];
+    }
+
     if (x.data.type === 'source') {
       if (x.data.source?.variableFrom === 'tool') {
         return [Object.values(x.data.source?.agent?.parameters || {})];
@@ -1160,12 +1212,14 @@ const useDelete = (value: AssistantYjs) => {
   const { deleteParameter } = useVariablesEditorOptions(value);
 
   const deleteUselessParameter = () => {
-    Object.values(value.parameters || {}).forEach((x) => {
-      const list = [FROM_API_PARAMETER, FROM_PARAMETER, FROM_API_PARAMETER];
-      if (x.data.from && list.includes(x.data.from) && !checkKeyParameterIsUsed({ value, key: x.data.key || '' })) {
-        deleteParameter(x.data);
-      }
-    });
+    Object.values(value.parameters || {})
+      .filter((i) => !i.data.hidden)
+      .forEach((x) => {
+        const list = [FROM_API_PARAMETER, FROM_PARAMETER, FROM_API_PARAMETER];
+        if (x.data.from && list.includes(x.data.from) && !checkKeyParameterIsUsed({ value, key: x.data.key || '' })) {
+          deleteParameter(x.data);
+        }
+      });
   };
 
   return {
@@ -1308,7 +1362,7 @@ function AgentParametersForm({
 
         <Box>
           {agent.parameters?.map((data) => {
-            if (!data?.key || data.type === 'source') return null;
+            if (!data?.key || data.type === 'source' || data.hidden) return null;
 
             const placeholder = data.placeholder?.replace(/([^\w]?)$/, '');
 
@@ -1341,7 +1395,7 @@ export function AuthorizeButton({ agent }: { agent: NonNullable<ReturnType<typeo
   const { t } = useLocaleContext();
 
   const authInputs = agent.parameters?.filter(
-    (i) => i.key && i.type === 'source' && i.source?.variableFrom === 'secret'
+    (i) => i.key && i.type === 'source' && i.source?.variableFrom === 'secret' && !i.hidden
   );
 
   const dialogState = usePopupState({ variant: 'dialog' });
@@ -1392,7 +1446,7 @@ function AuthorizeParametersFormDialog({
   const { t } = useLocaleContext();
 
   const authInputs = agent.parameters?.filter(
-    (i) => i.key && i.type === 'source' && i.source?.variableFrom === 'secret'
+    (i) => i.key && i.type === 'source' && i.source?.variableFrom === 'secret' && !i.hidden
   );
 
   const form = useForm();
@@ -1498,8 +1552,14 @@ function PopperButton({
           <Paper sx={{ p: 0, minWidth: 140, maxWidth: 320, maxHeight: '80vh', overflow: 'auto' }}>
             <Stack gap={2}>
               <List>
+                <MenuItem onClick={() => (parameter.hidden = !parameter.hidden)}>
+                  {parameter.hidden ? t('activeParameterTip') : t('hideParameterTip')}
+                </MenuItem>
+
                 {!(parameter.from === FROM_PARAMETER || parameter.from === FROM_KNOWLEDGE_PARAMETER) && (
-                  <MenuItem onClick={dialogState.open}>{t('setting')}</MenuItem>
+                  <MenuItem onClick={dialogState.open} disabled={Boolean(parameter.hidden)}>
+                    {t('setting')}
+                  </MenuItem>
                 )}
                 <MenuItem sx={{ color: '#E11D48', fontSize: 13 }} onClick={onDelete}>
                   {t('delete')}