+++ blocklets/ai-studio/src/components/file-editor/output/OpeningQuestionsSettings.tsx
@@ -102,7 +102,7 @@ function EntryItemView({
   const doc = (getYjsValue(item) as Map<any>).doc!;
 
   const parameters = sortBy(Object.values(assistant.parameters ?? {}), (i) => i.index).filter(
-    (i): i is typeof i & { data: { key: string } } => !!i.data.key
+    (i): i is typeof i & { data: { key: string; hidden?: boolean } } => !!i.data.key && !i.data.hidden
   );
 
   return (